Separate Ghc from Ghci
