+++ lib/autocomplete-provider.js
@@ -1,13 +1,13 @@
 'use babel';
 
-const Ghci = require('./ghci');
+const Ghc = require('./ghc');
 
 export default class AutocompleteProvider {
 
   constructor() {
     this.selector = '.source.tidalcycles';
     if (atom.config.get('tidalcycles.autocomplete')) {
-      Ghci.exec(`:browse Sound.Tidal.Context`, (output) => {
+      Ghc.exec(`:browse Sound.Tidal.Context`, (output) => {
         this.suggestions = this.parse(output)
       })
     } else {
