+++ lib/boot-tidal.js
@@ -2,8 +2,7 @@
 
 const path = require('path');
 const fs = require('fs');
-const child_process = require('child_process');
-const Ghci = require('./ghci')
+const Ghc = require('./ghc')
 
 const defaultBootFileName = 'BootTidal.hs';
 const defaultBootFilePath = __dirname + path.sep + defaultBootFileName;
@@ -29,7 +28,7 @@ export default class BootTidal {
 
     if (fs.existsSync(currentDirectoryPath)) return currentDirectoryPath;
 
-    let tidalBootPath = path.join(Ghci.tidalDataDir(), defaultBootFileName)
+    let tidalBootPath = path.join(Ghc.tidalDataDir(), defaultBootFileName)
     if (fs.existsSync(tidalBootPath)) {
       return tidalBootPath
     } else {
