+++ lib/ghci.js
@@ -1,53 +1,8 @@
 'use babel';
 
-const child_process = require('child_process');
-const { exec, spawn } = require('child_process');
-const path = require('path');
 const EventEmitter = require('events');
-const ghciPathProperty = 'tidalcycles.ghciPath'
 
 export default class Ghci extends EventEmitter {
-
-  static spawn() {
-    return new Ghci(
-      spawn(Ghci.commandPath('ghci'), [], { shell: true })
-    )
-  }
-
-  static commandPath(name) {
-    let propertyValue = Ghci.ghciPathProperty()
-    if (propertyValue) {
-      return path.basename(propertyValue) === 'ghci'
-        ? path.join(path.parse(propertyValue).dir, name)
-        : path.join(propertyValue, name)
-    } else {
-      return name
-    }
-  }
-
-  static ghciPathProperty() {
-    return atom.config.get(ghciPathProperty)
-  }
-
-  static exec(command, callback) {
-    let ghciPath = Ghci.commandPath('ghci')
-    exec(`echo "${command}" | ${ghciPath}`, (error, stdout) => {
-      if (error) {
-        console.error(`exec error: ${error}`);
-        return;
-      }
-
-      callback(stdout)
-    })
-  }
-
-  static tidalDataDir() {
-    let dataDir = child_process
-      .execSync(`${Ghci.commandPath('ghc-pkg')} field tidal data-dir`)
-      .toString().trim()
-
-    return dataDir.substring(dataDir.indexOf(' ') + 1)
-  }
 
   constructor(process) {
     super()
