+++ lib/repl.js
@@ -1,7 +1,7 @@
 'use babel';
 
 var fs = require('fs');
-const Ghci = require('./ghci')
+const Ghc = require('./ghc')
 const BootTidal = require('./boot-tidal')
 var CONST_LINE = 'line'
 var CONST_MULTI_LINE = 'multi_line'
@@ -51,7 +51,7 @@ export default class REPL {
 
     start() {
       this.consoleView.initUI();
-      this.ghci = Ghci.spawn()
+      this.ghci = Ghc.interactive()
         .on('stderr', data => { this.processStdErr(data) })
         .on('stdout', data => { this.processStdOut(data) })
